+++ girder/api/v1/notification.py
@@ -24,7 +24,10 @@ from datetime import datetime
 
 from ..describe import Description, autoDescribeRoute
 from ..rest import Resource, setResponseHeader
+from girder.constants import SettingKey
+from girder.exceptions import RestException
 from girder.models.notification import Notification as NotificationModel
+from girder.models.setting import Setting
 from girder.utility import JsonEncoder
 from girder.api import access
 
@@ -71,8 +74,12 @@ class Notification(Resource):
         .produces('text/event-stream')
         .errorResponse()
         .errorResponse('You are not logged in.', 403)
+        .errorResponse('The notification stream is not enabled.', 503)
     )
     def stream(self, timeout, params):
+        if not Setting().get(SettingKey.ENABLE_NOTIFICATION_STREAM):
+            raise RestException('The notification stream is not enabled.', code=503)
+
         user, token = self.getCurrentUser(returnToken=True)
 
         setResponseHeader('Content-Type', 'text/event-stream')