+++ girder/constants.py
@@ -174,6 +174,7 @@ class SettingKey(object):
     EMAIL_VERIFICATION = 'core.email_verification'
     ENABLE_PASSWORD_LOGIN = 'core.enable_password_login'
     GIRDER_MOUNT_INFORMATION = 'core.girder_mount_information'
+    ENABLE_NOTIFICATION_STREAM = 'core.enable_notification_stream'
     PLUGINS_ENABLED = 'core.plugins_enabled'
     REGISTRATION_POLICY = 'core.registration_policy'
     ROUTE_TABLE = 'core.route_table'
@@ -216,6 +217,7 @@ class SettingDefault(object):
         SettingKey.EMAIL_VERIFICATION: 'disabled',
         SettingKey.EMAIL_FROM_ADDRESS: 'Girder <no-reply@girder.org>',
         SettingKey.ENABLE_PASSWORD_LOGIN: True,
+        SettingKey.ENABLE_NOTIFICATION_STREAM: True,
         SettingKey.PLUGINS_ENABLED: [],
         SettingKey.REGISTRATION_POLICY: 'open',
         SettingKey.SMTP_HOST: 'localhost',